chore(create-playwright): nits and cleanups
