+++ packages/create-playwright/assets/playwright.config.ts
@@ -35,7 +35,6 @@ const config: PlaywrightTestConfig = {
       name: 'Desktop Chrome',
       use: {
         ...devices['Desktop Chrome'],
-
       },
     },
     {
